+++ ixmp/backend/jdbc.py
@@ -18,7 +18,7 @@ import pandas as pd
 from ixmp import config
 from ixmp.core import Scenario
 from ixmp.utils import as_str_list, filtered, islistable
-from . import FIELDS
+from . import FIELDS, ItemType
 from .base import CachingBackend
 
 
@@ -293,6 +293,110 @@ class JDBCBackend(CachingBackend):
 
     def get_units(self):
         return to_pylist(self.jobj.getUnitList())
+
+    def read_file(self, path, item_type: ItemType, **kwargs):
+        """Read Platform, TimeSeries, or Scenario data from file.
+
+        JDBCBackend supports reading from:
+
+        - ``path='*.gdx', item_type=ItemType.MODEL``. The keyword arguments
+          `check_solution`, `comment`, `equ_list`, and `var_list` are
+          **required**.
+
+        Other parameters
+        ----------------
+        check_solution : bool
+            If True, raise an exception if the GAMS solver did not reach
+            optimality. (Only for MESSAGE-scheme Scenarios.)
+        comment : str
+            Comment added to Scenario when importing the solution.
+        equ_list : list of str
+            Equations to be imported.
+        var_list : list of str
+            Variables to be imported.
+        filters : dict of dict of str
+            Restrict items read.
+
+        See also
+        --------
+        .Backend.read_file
+        """
+        ts, filters = self._handle_rw_filters(kwargs.pop('filters', {}))
+        if path.suffix == '.gdx' and item_type is ItemType.MODEL:
+            kw = {'check_solution', 'comment', 'equ_list', 'var_list'}
+
+            if not isinstance(ts, Scenario):
+                raise ValueError('read from GDX requires a Scenario object')
+            elif set(kwargs.keys()) != kw:
+                raise ValueError(('keyword arguments {} do not match required '
+                                  '{}').format(kwargs.keys(), kw))
+
+            args = (
+                str(path.parent),
+                path.name,
+                kwargs.pop('comment'),
+                to_jlist2(kwargs.pop('var_list')),
+                to_jlist2(kwargs.pop('equ_list')),
+                kwargs.pop('check_solution'),
+            )
+
+            if len(kwargs):
+                raise ValueError('extra keyword arguments {}'.format(kwargs))
+
+            self.jindex[ts].readSolutionFromGDX(*args)
+
+            self.cache_invalidate(ts)
+        else:
+            raise NotImplementedError(path, item_type)
+
+    def write_file(self, path, item_type: ItemType, **kwargs):
+        """Write Platform, TimeSeries, or Scenario data to file.
+
+        JDBCBackend supports writing to:
+
+        - ``path='*.gdx', item_type=ItemType.SET | ItemType.PAR``.
+        - ``path='*.csv', item_type=TS``. The `default` keyword argument is
+          **required**.
+
+        Other parameters
+        ----------------
+        filters : dict of dict of str
+            Restrict items written.
+            Following filters may be used:
+            - model : str
+            - scenario : str
+            - variable : list of str
+            - default : bool
+              If :obj:`True`, only data from TimeSeries versions with
+              :meth:`set_default` are written.
+
+        See also
+        --------
+        .Backend.write_file
+        """
+        ts, filters = self._handle_rw_filters(kwargs.pop('filters', {}))
+        if path.suffix == '.gdx' and item_type is ItemType.SET | ItemType.PAR:
+            if len(filters):
+                raise NotImplementedError('write to GDX with filters')
+            elif not isinstance(ts, Scenario):
+                raise ValueError('write to GDX requires a Scenario object')
+
+            # include_var_equ=False -> do not include variables/equations in
+            # GDX
+            self.jindex[ts].toGDX(str(path.parent), path.name, False)
+        elif path.suffix == '.csv' and item_type is ItemType.TS:
+            model = filters.pop('model')
+            scenario = filters.pop('scenario')
+            variables = filters.pop('variable')
+            default = filters.pop('default')
+
+            scen_list = self.jobj.getScenarioList(default, model, scenario)
+            run_ids = [s['run_id'] for s in scen_list]
+            self.jobj.exportTimeseriesData(to_jlist2(run_ids),
+                                           to_jlist2(variables),
+                                           str(path))
+        else:
+            raise NotImplementedError
 
     # Timeseries methods
 
@@ -701,32 +805,6 @@ class JDBCBackend(CachingBackend):
         self.jindex[ms].addCatEle(name, cat, to_jlist2(keys), is_unique)
 
     # Helpers; not part of the Backend interface
-
-    def write_gdx(self, s, path):
-        """Write the Scenario to a GDX file at *path*."""
-        # include_var_equ=False -> do not include variables/equations in GDX
-        self.jindex[s].toGDX(str(path.parent), path.name, False)
-
-    def read_gdx(self, s, path, check_solution, comment, equ_list, var_list):
-        """Read the Scenario from a GDX file at *path*.
-
-        Parameters
-        ----------
-        check_solution : bool
-            If True, raise an exception if the GAMS solver did not reach
-            optimality. (Only for MESSAGE-scheme Scenarios.)
-        comment : str
-            Comment added to Scenario when importing the solution.
-        equ_list : list of str
-            Equations to be imported.
-        var_list : list of str
-            Variables to be imported.
-        """
-        self.jindex[s].readSolutionFromGDX(
-            str(path.parent), path.name, comment, to_jlist2(var_list),
-            to_jlist2(equ_list), check_solution)
-
-        self.cache_invalidate(s)
 
     def _get_item(self, s, ix_type, name, load=True):
         """Return the Java object for item *name* of *ix_type*.